+++ CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 
 Upcoming
 --------
+* added `json` and `pickle` output file support in `run.py`
 
 2023-07-13 (0.0.1)
 -----------------